Move the utility to IREE::LinalgExt
